[3.x] GLTF: Organize structures and extensions into subfolders
