+++ modules/gltf/extensions/SCsub
@@ -7,5 +7,3 @@ env_gltf = env_modules.Clone()
 
 # Godot source files
 env_gltf.add_source_files(env.modules_sources, "*.cpp")
-env_gltf.add_source_files(env.modules_sources, "structures/*.cpp")
-SConscript("extensions/SCsub")
