+++ modules/gltf/extensions/gltf_light.cpp
