+++ modules/gltf/gltf_state.h
@@ -31,18 +31,18 @@
 #ifndef GLTF_STATE_H
 #define GLTF_STATE_H
 
-#include "gltf_accessor.h"
-#include "gltf_animation.h"
-#include "gltf_buffer_view.h"
-#include "gltf_camera.h"
-#include "gltf_light.h"
-#include "gltf_mesh.h"
-#include "gltf_node.h"
-#include "gltf_skeleton.h"
-#include "gltf_skin.h"
+#include "extensions/gltf_light.h"
 #include "gltf_template_convert.h"
-#include "gltf_texture.h"
-#include "gltf_texture_sampler.h"
+#include "structures/gltf_accessor.h"
+#include "structures/gltf_animation.h"
+#include "structures/gltf_buffer_view.h"
+#include "structures/gltf_camera.h"
+#include "structures/gltf_mesh.h"
+#include "structures/gltf_node.h"
+#include "structures/gltf_skeleton.h"
+#include "structures/gltf_skin.h"
+#include "structures/gltf_texture.h"
+#include "structures/gltf_texture_sampler.h"
 
 class GLTFState : public Resource {
 	GDCLASS(GLTFState, Resource);
