+++ modules/gltf/structures/gltf_mesh.h
