+++ modules/gltf/structures/gltf_node.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  gltf_texture.h                                                       */
+/*  gltf_node.h                                                          */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,27 +28,75 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef GLTF_TEXTURE_H
-#define GLTF_TEXTURE_H
+#ifndef GLTF_NODE_H
+#define GLTF_NODE_H
 
+#include "../gltf_defines.h"
 #include "core/resource.h"
-#include "gltf_defines.h"
 
-class GLTFTexture : public Resource {
-	GDCLASS(GLTFTexture, Resource);
+class GLTFNode : public Resource {
+	GDCLASS(GLTFNode, Resource);
+	friend class GLTFDocument;
+	friend class PackedSceneGLTF;
 
 private:
-	GLTFImageIndex src_image = 0;
-	GLTFTextureSamplerIndex sampler = -1;
+	// matrices need to be transformed to this
+	GLTFNodeIndex parent = -1;
+	int height = -1;
+	Transform xform;
+	GLTFMeshIndex mesh = -1;
+	GLTFCameraIndex camera = -1;
+	GLTFSkinIndex skin = -1;
+	GLTFSkeletonIndex skeleton = -1;
+	bool joint = false;
+	Vector3 translation;
+	Quat rotation;
+	Vector3 scale = Vector3(1, 1, 1);
+	Vector<int> children;
+	GLTFLightIndex light = -1;
 
 protected:
 	static void _bind_methods();
 
 public:
-	GLTFImageIndex get_src_image() const;
-	void set_src_image(GLTFImageIndex val);
-	GLTFTextureSamplerIndex get_sampler() const;
-	void set_sampler(GLTFTextureSamplerIndex val);
+	GLTFNodeIndex get_parent();
+	void set_parent(GLTFNodeIndex p_parent);
+
+	int get_height();
+	void set_height(int p_height);
+
+	Transform get_xform();
+	void set_xform(Transform p_xform);
+
+	GLTFMeshIndex get_mesh();
+	void set_mesh(GLTFMeshIndex p_mesh);
+
+	GLTFCameraIndex get_camera();
+	void set_camera(GLTFCameraIndex p_camera);
+
+	GLTFSkinIndex get_skin();
+	void set_skin(GLTFSkinIndex p_skin);
+
+	GLTFSkeletonIndex get_skeleton();
+	void set_skeleton(GLTFSkeletonIndex p_skeleton);
+
+	bool get_joint();
+	void set_joint(bool p_joint);
+
+	Vector3 get_translation();
+	void set_translation(Vector3 p_translation);
+
+	Quat get_rotation();
+	void set_rotation(Quat p_rotation);
+
+	Vector3 get_scale();
+	void set_scale(Vector3 p_scale);
+
+	Vector<int> get_children();
+	void set_children(Vector<int> p_children);
+
+	GLTFLightIndex get_light();
+	void set_light(GLTFLightIndex p_light);
 };
 
-#endif // GLTF_TEXTURE_H
+#endif // GLTF_NODE_H
