+++ modules/gltf/structures/gltf_texture.h
@@ -1,5 +1,5 @@
 /*************************************************************************/
-/*  gltf_skeleton.h                                                      */
+/*  gltf_texture.h                                                       */
 /*************************************************************************/
 /*                       This file is part of:                           */
 /*                           GODOT ENGINE                                */
@@ -28,55 +28,27 @@
 /* SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.                */
 /*************************************************************************/
 
-#ifndef GLTF_SKELETON_H
-#define GLTF_SKELETON_H
+#ifndef GLTF_TEXTURE_H
+#define GLTF_TEXTURE_H
 
+#include "../gltf_defines.h"
 #include "core/resource.h"
-#include "gltf_defines.h"
 
-class GLTFSkeleton : public Resource {
-	GDCLASS(GLTFSkeleton, Resource);
-	friend class GLTFDocument;
+class GLTFTexture : public Resource {
+	GDCLASS(GLTFTexture, Resource);
 
 private:
-	// The *synthesized* skeletons joints
-	PoolVector<GLTFNodeIndex> joints;
-
-	// The roots of the skeleton. If there are multiple, each root must have the
-	// same parent (ie roots are siblings)
-	PoolVector<GLTFNodeIndex> roots;
-
-	// The created Skeleton for the scene
-	Skeleton *godot_skeleton = nullptr;
-
-	// Set of unique bone names for the skeleton
-	Set<String> unique_names;
-
-	Map<int32_t, GLTFNodeIndex> godot_bone_node;
-
-	PoolVector<BoneAttachment *> bone_attachments;
+	GLTFImageIndex src_image = 0;
+	GLTFTextureSamplerIndex sampler = -1;
 
 protected:
 	static void _bind_methods();
 
 public:
-	PoolVector<GLTFNodeIndex> get_joints();
-	void set_joints(PoolVector<GLTFNodeIndex> p_joints);
-
-	PoolVector<GLTFNodeIndex> get_roots();
-	void set_roots(PoolVector<GLTFNodeIndex> p_roots);
-
-	Skeleton *get_godot_skeleton();
-
-	Array get_unique_names();
-	void set_unique_names(Array p_unique_names);
-
-	Dictionary get_godot_bone_node();
-	void set_godot_bone_node(Dictionary p_indict);
-
-	BoneAttachment *get_bone_attachment(int idx);
-
-	int32_t get_bone_attachment_count();
+	GLTFImageIndex get_src_image() const;
+	void set_src_image(GLTFImageIndex val);
+	GLTFTextureSamplerIndex get_sampler() const;
+	void set_sampler(GLTFTextureSamplerIndex val);
 };
 
-#endif // GLTF_SKELETON_H
+#endif // GLTF_TEXTURE_H
